+++ purchase_addresses/__openerp__.py
@@ -29,15 +29,16 @@
  'images': [],
  'website': "http://www.camptocamp.com",
  'depends': ['purchase',
-             'stock_addresses'
+             'stock_addresses',
+             'sale_addresses',
              ],
  'demo': [],
  'data': ['view/purchase.xml',
           ],
- 'test': ['setup_user.yml',
-          'setup_product.yml',
-          'setup_dropshipping.xml',
-          'test_propagate_address.yml',
+ 'test': ['test/setup_user.yml',
+          'test/setup_product.yml',
+          'test/setup_dropshipping.xml',
+          'test/test_propagate_address.yml',
           ],
  'auto_install': False,
  'installable': True,
